refactor: add HeaderGroup, remove conditional logic from App
