+++ src/components/App/App.js
@@ -12,13 +12,9 @@ import SavedNews from "../SavedNews/SavedNews";
 import RegisterSuccessModal from "../RegisterSuccessModal/RegisterSuccessModal";
 import LoginRegisterModal from "../LoginRegisterModal/LoginRegisterModal";
 import Footer from "../Footer/Footer";
-import { useLocation } from "react-router-dom/cjs/react-router-dom.min";
-import HomeImageBackground from "../HomeImageBackground/HomeImageBackground";
-import Header from "../Header/Header";
-import SearchForm from '../SearchForm/SearchForm';
+import HeaderGroup from "../HeaderGroup/HeaderGroup";
 
 function App() {
-  const location = useLocation();
   const [loggedIn, setLoggedIn] = useState(false);
   const [isSearching, setIsSearching] = useState(false);
   const [isSearchResultsOpen, setIsSearchResultsOpen] = useState(false);
@@ -281,50 +277,37 @@ function App() {
     <Wrapper>
       <GlobalStyle />
       <LoggedInContext.Provider value={loggedIn}>
-      {location.pathname === '/' ?
-        (
-          <HomeImageBackground>
-            <Header
-              openLoginWindow={openLoginWindow}
-              handleLogOut={handleLogOut}
-            />
-            <SearchForm handleSearchSubmit={handleSearchSubmit}/>
-          </HomeImageBackground>
-        )
-      : 
-        (
-          <Header
-            openLoginWindow={openLoginWindow}
-            handleLogOut={handleLogOut}
-          />
-        )
-      }
-      <Switch>
-          <ProtectedRoute exact path='/saved-news'>
-            <SavedNews 
-              openLoginWindow={openLoginWindow}
-              handleLogOut={handleLogOut}
-              articleCount={savedArticles.length}
-              displayArticles={savedArticles}
-              cardIconClick={updateSaved}
-            />
-          </ProtectedRoute>
-          <Route exact path='/'>
-            <Main 
-              openLoginWindow={openLoginWindow}
-              handleLogOut={handleLogOut}
-              isSearching={isSearching}
-              isSearchResultsOpen={isSearchResultsOpen}
-              noSearchResults={noSearchResults}
-              isErrorOpen={isErrorOpen}
-              handleSearchSubmit={handleSearchSubmit}
-              displayArticles={displayArticles}
-              handleShowMoreClick={handleShowMoreClick}
-              moreArticles={moreArticles}
-              cardIconClick={updateSaved}
-            />
-          </Route>
-      </Switch>
+        <HeaderGroup 
+          openLoginWindow={openLoginWindow}
+          handleLogOut={handleLogOut}
+          handleSearchSubmit={handleSearchSubmit}
+        />
+        <Switch>
+            <ProtectedRoute exact path='/saved-news'>
+              <SavedNews 
+                openLoginWindow={openLoginWindow}
+                handleLogOut={handleLogOut}
+                articleCount={savedArticles.length}
+                displayArticles={savedArticles}
+                cardIconClick={updateSaved}
+              />
+            </ProtectedRoute>
+            <Route exact path='/'>
+              <Main 
+                openLoginWindow={openLoginWindow}
+                handleLogOut={handleLogOut}
+                isSearching={isSearching}
+                isSearchResultsOpen={isSearchResultsOpen}
+                noSearchResults={noSearchResults}
+                isErrorOpen={isErrorOpen}
+                handleSearchSubmit={handleSearchSubmit}
+                displayArticles={displayArticles}
+                handleShowMoreClick={handleShowMoreClick}
+                moreArticles={moreArticles}
+                cardIconClick={updateSaved}
+              />
+            </Route>
+        </Switch>
       </LoggedInContext.Provider>
       <Footer />
       <RegisterSuccessModal
