--- src/components/HomeImageBackground/HomeImageBackground.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import { ImageWrapper } from './styledHomeImageBackground'
-
-export default function HomeImageBackground({ children }) {
-  return (
-    <ImageWrapper>
-      { children }
-    </ImageWrapper>
-  )
-}
